Update latex_misc subdirectory to different remote
The latex_misc subdirectory is a local copy of another repo.
This commit changes the remote repo associated with this directory
and replaces it with the one on github. The new subdirectory
version number is v0.2.2. The bibtex records in references.bib
should be more up to date, but the more important addition is
`version_git.sh` which will automatically generate LaTeX macros
containing git commit information.